integration hbase rest into hbase
