+++ hadoop-mini-clusters-hbase/src/test/java/com/github/sakserv/minicluster/impl/HbaseLocalClusterTest.java
@@ -19,6 +19,7 @@ import static org.junit.Assert.assertTrue;
 import java.io.IOException;
 
 import org.apache.hadoop.conf.Configuration;
+import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
@@ -47,10 +48,10 @@ public class HbaseLocalClusterTest {
     @Rule
     public ExpectedException exception = ExpectedException.none();
 
-    private static HbaseLocalCluster hbaseLocalCluster;
+    private HbaseLocalCluster hbaseLocalCluster;
 
-    @BeforeClass
-    public static void setUp(){
+    @Before
+    public void setUp(){
         hbaseLocalCluster = new HbaseLocalCluster.Builder()
                 .setHbaseMasterPort(
                         Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_PORT_KEY)))
@@ -65,6 +66,19 @@ public class HbaseLocalClusterTest {
                 .setHbaseWalReplicationEnabled(
                         Boolean.parseBoolean(propertyParser.getProperty(ConfigVars.HBASE_WAL_REPLICATION_ENABLED_KEY)))
                 .setHbaseConfiguration(new Configuration())
+                .activeRestGateway()
+                    .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
+                    .setHbaseRestPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
+                    .setHbaseRestInfoPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
+                    .setHbaseRestReadOnly(
+                            Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
+                    .setHbaseRestThreadMax(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
+                    .setHbaseRestThreadMin(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
+                    .build()
                 .build();
     }
 
@@ -89,6 +103,19 @@ public class HbaseLocalClusterTest {
                 .setHbaseWalReplicationEnabled(
                         Boolean.parseBoolean(propertyParser.getProperty(ConfigVars.HBASE_WAL_REPLICATION_ENABLED_KEY)))
                 .setHbaseConfiguration(new Configuration())
+                .activeRestGateway()
+                    .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
+                    .setHbaseRestPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
+                    .setHbaseRestInfoPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
+                    .setHbaseRestReadOnly(
+                            Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
+                    .setHbaseRestThreadMax(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
+                    .setHbaseRestThreadMin(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
+                    .build()
                 .build();
     }
 
@@ -113,6 +140,19 @@ public class HbaseLocalClusterTest {
                 .setHbaseWalReplicationEnabled(
                         Boolean.parseBoolean(propertyParser.getProperty(ConfigVars.HBASE_WAL_REPLICATION_ENABLED_KEY)))
                 .setHbaseConfiguration(new Configuration())
+                .activeRestGateway()
+                    .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
+                    .setHbaseRestPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
+                    .setHbaseRestInfoPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
+                    .setHbaseRestReadOnly(
+                            Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
+                    .setHbaseRestThreadMax(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
+                    .setHbaseRestThreadMin(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
+                    .build()
                 .build();
     }
 
@@ -137,6 +177,19 @@ public class HbaseLocalClusterTest {
                 .setHbaseWalReplicationEnabled(
                         Boolean.parseBoolean(propertyParser.getProperty(ConfigVars.HBASE_WAL_REPLICATION_ENABLED_KEY)))
                 .setHbaseConfiguration(new Configuration())
+                .activeRestGateway()
+                    .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
+                    .setHbaseRestPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
+                    .setHbaseRestInfoPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
+                    .setHbaseRestReadOnly(
+                            Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
+                    .setHbaseRestThreadMax(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
+                    .setHbaseRestThreadMin(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
+                    .build()
                 .build();
     }
 
@@ -161,6 +214,19 @@ public class HbaseLocalClusterTest {
                 .setHbaseWalReplicationEnabled(
                         Boolean.parseBoolean(propertyParser.getProperty(ConfigVars.HBASE_WAL_REPLICATION_ENABLED_KEY)))
                 .setHbaseConfiguration(new Configuration())
+                .activeRestGateway()
+                    .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
+                    .setHbaseRestPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
+                    .setHbaseRestInfoPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
+                    .setHbaseRestReadOnly(
+                            Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
+                    .setHbaseRestThreadMax(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
+                    .setHbaseRestThreadMin(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
+                    .build()
                 .build();
     }
 
@@ -186,6 +252,19 @@ public class HbaseLocalClusterTest {
                 .setHbaseWalReplicationEnabled(
                         Boolean.parseBoolean(propertyParser.getProperty(ConfigVars.HBASE_WAL_REPLICATION_ENABLED_KEY)))
                 .setHbaseConfiguration(new Configuration())
+                .activeRestGateway()
+                    .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
+                    .setHbaseRestPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
+                    .setHbaseRestInfoPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
+                    .setHbaseRestReadOnly(
+                            Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
+                    .setHbaseRestThreadMax(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
+                    .setHbaseRestThreadMin(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
+                    .build()
                 .build();
     }
 
@@ -211,6 +290,19 @@ public class HbaseLocalClusterTest {
                 .setHbaseWalReplicationEnabled(
                         Boolean.parseBoolean(propertyParser.getProperty(ConfigVars.HBASE_WAL_REPLICATION_ENABLED_KEY)))
                 .setHbaseConfiguration(new Configuration())
+                .activeRestGateway()
+                    .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
+                    .setHbaseRestPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
+                    .setHbaseRestInfoPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
+                    .setHbaseRestReadOnly(
+                            Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
+                    .setHbaseRestThreadMax(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
+                    .setHbaseRestThreadMin(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
+                    .build()
                 .build();
     }
 
@@ -236,6 +328,19 @@ public class HbaseLocalClusterTest {
                 .setHbaseWalReplicationEnabled(
                         Boolean.parseBoolean(propertyParser.getProperty(ConfigVars.HBASE_WAL_REPLICATION_ENABLED_KEY)))
                 .setHbaseConfiguration(new Configuration())
+                .activeRestGateway()
+                    .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
+                    .setHbaseRestPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
+                    .setHbaseRestInfoPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
+                    .setHbaseRestReadOnly(
+                            Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
+                    .setHbaseRestThreadMax(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
+                    .setHbaseRestThreadMin(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
+                    .build()
                 .build();
     }
 
@@ -260,6 +365,19 @@ public class HbaseLocalClusterTest {
                 .setZookeeperConnectionString(propertyParser.getProperty(ConfigVars.ZOOKEEPER_CONNECTION_STRING_KEY))
                 .setZookeeperZnodeParent(propertyParser.getProperty(ConfigVars.HBASE_ZNODE_PARENT_KEY))
                 .setHbaseConfiguration(new Configuration())
+                .activeRestGateway()
+                    .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
+                    .setHbaseRestPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
+                    .setHbaseRestInfoPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
+                    .setHbaseRestReadOnly(
+                            Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
+                    .setHbaseRestThreadMax(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
+                    .setHbaseRestThreadMin(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
+                    .build()
                 .build();
     }
 
@@ -285,6 +403,88 @@ public class HbaseLocalClusterTest {
                 .setZookeeperZnodeParent(propertyParser.getProperty(ConfigVars.HBASE_ZNODE_PARENT_KEY))
                 .setHbaseWalReplicationEnabled(
                         Boolean.parseBoolean(propertyParser.getProperty(ConfigVars.HBASE_WAL_REPLICATION_ENABLED_KEY)))
+                .activeRestGateway()
+                    .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
+                    .setHbaseRestPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
+                    .setHbaseRestInfoPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
+                    .setHbaseRestReadOnly(
+                            Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
+                    .setHbaseRestThreadMax(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
+                    .setHbaseRestThreadMin(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
+                    .build()
                 .build();
     }
+
+
+    @Test
+    public void testHbaseRestPort() {
+        assertEquals(Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)),
+                (int) hbaseLocalCluster.getHbaseRestLocalCluster().getHbaseRestPort());
+    }
+
+    @Test
+    public void testMissingHbaseRestPort() {
+        exception.expect(IllegalArgumentException.class);
+        hbaseLocalCluster = new HbaseLocalCluster.Builder()
+                .setHbaseMasterPort(
+                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_PORT_KEY)))
+                .setHbaseMasterInfoPort(
+                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_INFO_PORT_KEY)))
+                .setNumRegionServers(
+                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_NUM_REGION_SERVERS_KEY)))
+                .setHbaseRootDir(propertyParser.getProperty(ConfigVars.HBASE_ROOT_DIR_KEY))
+                .setZookeeperPort(
+                        Integer.parseInt(propertyParser.getProperty(ConfigVars.ZOOKEEPER_PORT_KEY)))
+                .setZookeeperConnectionString(propertyParser.getProperty(ConfigVars.ZOOKEEPER_CONNECTION_STRING_KEY))
+                .setZookeeperZnodeParent(propertyParser.getProperty(ConfigVars.HBASE_ZNODE_PARENT_KEY))
+                .activeRestGateway()
+                    .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
+                    .setHbaseRestPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
+                    .setHbaseRestInfoPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
+                    .setHbaseRestReadOnly(
+                            Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
+                    .setHbaseRestThreadMax(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
+                    .setHbaseRestThreadMin(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
+                    .build()
+                .build();
+    }
+
+    @Test
+    public void testHbaseRestInfoPort() {
+        assertEquals(Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)),
+                (int) hbaseLocalCluster.getHbaseRestLocalCluster().getHbaseRestInfoPort());
+    }
+
+    @Test
+    public void testHbaseRestHost() {
+        assertEquals(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY),
+                hbaseLocalCluster.getHbaseRestLocalCluster().getHbaseRestHost());
+    }
+
+    @Test
+    public void testHbaseRestReadOnly() {
+        assertEquals(Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)),
+                hbaseLocalCluster.getHbaseRestLocalCluster().getHbaseRestReadOnly());
+    }
+
+    @Test
+    public void testHbaseRestThreadMax() {
+        assertEquals(Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)),
+                hbaseLocalCluster.getHbaseRestLocalCluster().getHbaseRestThreadMax());
+    }
+
+    @Test
+    public void testHbaseRestThreadMin() {
+        assertEquals(Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)),
+                hbaseLocalCluster.getHbaseRestLocalCluster().getHbaseRestThreadMin());
+    }
+
 }
