--- hadoop-mini-clusters-hbase/src/test/java/com/github/sakserv/minicluster/impl/HbaseRestLocalClusterTest.java
@@ -1,370 +1,0 @@
-/*
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-package com.github.sakserv.minicluster.impl;
-
-import com.github.sakserv.minicluster.config.ConfigVars;
-import com.github.sakserv.propertyparser.PropertyParser;
-import org.junit.BeforeClass;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.ExpectedException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.IOException;
-
-import static org.junit.Assert.assertEquals;
-
-public class HbaseRestLocalClusterTest {
-    // Logger
-    private static final Logger LOG = LoggerFactory.getLogger(HbaseRestLocalClusterTest.class);
-
-    // Setup the property parser
-    private static PropertyParser propertyParser;
-
-    static {
-        try {
-            propertyParser = new PropertyParser(ConfigVars.DEFAULT_PROPS_FILE);
-            propertyParser.parsePropsFile();
-        } catch (IOException e) {
-            LOG.error("Unable to load property file: {}", propertyParser.getProperty(ConfigVars.DEFAULT_PROPS_FILE));
-        }
-    }
-
-    @Rule
-    public ExpectedException exception = ExpectedException.none();
-
-    private static HbaseRestLocalCluster hbaseRestLocalCluster;
-
-    @BeforeClass
-    public static void setUp() {
-        hbaseRestLocalCluster = new HbaseRestLocalCluster.Builder()
-                .setHbaseMasterPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_PORT_KEY)))
-                .setHbaseMasterInfoPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_INFO_PORT_KEY)))
-                .setNumRegionServers(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_NUM_REGION_SERVERS_KEY)))
-                .setHbaseRootDir(propertyParser.getProperty(ConfigVars.HBASE_ROOT_DIR_KEY))
-                .setZookeeperPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.ZOOKEEPER_PORT_KEY)))
-                .setZookeeperConnectionString(propertyParser.getProperty(ConfigVars.ZOOKEEPER_CONNECTION_STRING_KEY))
-                .setZookeeperZnodeParent(propertyParser.getProperty(ConfigVars.HBASE_ZNODE_PARENT_KEY))
-                .setHbaseRestPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
-                .setHbaseRestInfoPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
-                .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
-                .setHbaseRestReadOnly(
-                        Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
-                .setHbaseRestThreadMax(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
-                .setHbaseRestThreadMin(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
-                .build();
-    }
-
-    @Test
-    public void testHbaseMasterPort() {
-        assertEquals(Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_PORT_KEY)),
-                (int) hbaseRestLocalCluster.getHbaseMasterPort());
-    }
-
-    @Test
-    public void testMissingHbaseMasterPort() {
-        exception.expect(IllegalArgumentException.class);
-        hbaseRestLocalCluster = new HbaseRestLocalCluster.Builder()
-                .setHbaseMasterInfoPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_INFO_PORT_KEY)))
-                .setNumRegionServers(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_NUM_REGION_SERVERS_KEY)))
-                .setHbaseRootDir(propertyParser.getProperty(ConfigVars.HBASE_ROOT_DIR_KEY))
-                .setZookeeperPort(Integer.parseInt(propertyParser.getProperty(ConfigVars.ZOOKEEPER_PORT_KEY)))
-                .setZookeeperConnectionString(propertyParser.getProperty(ConfigVars.ZOOKEEPER_CONNECTION_STRING_KEY))
-                .setZookeeperZnodeParent(propertyParser.getProperty(ConfigVars.HBASE_ZNODE_PARENT_KEY))
-                .setHbaseRestPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
-                .setHbaseRestInfoPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
-                .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
-                .setHbaseRestReadOnly(
-                        Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
-                .setHbaseRestThreadMax(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
-                .setHbaseRestThreadMin(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
-                .build();
-    }
-
-    @Test
-    public void testHbaseMasterInfoPort() {
-        assertEquals(Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_INFO_PORT_KEY)),
-                (int) hbaseRestLocalCluster.getHbaseMasterInfoPort());
-    }
-
-    @Test
-    public void testMissingHbaseMasterInfoPort() {
-        exception.expect(IllegalArgumentException.class);
-        hbaseRestLocalCluster = new HbaseRestLocalCluster.Builder()
-                .setHbaseMasterPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_PORT_KEY)))
-                .setNumRegionServers(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_NUM_REGION_SERVERS_KEY)))
-                .setHbaseRootDir(propertyParser.getProperty(ConfigVars.HBASE_ROOT_DIR_KEY))
-                .setZookeeperPort(Integer.parseInt(propertyParser.getProperty(ConfigVars.ZOOKEEPER_PORT_KEY)))
-                .setZookeeperConnectionString(propertyParser.getProperty(ConfigVars.ZOOKEEPER_CONNECTION_STRING_KEY))
-                .setZookeeperZnodeParent(propertyParser.getProperty(ConfigVars.HBASE_ZNODE_PARENT_KEY))
-                .setHbaseRestPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
-                .setHbaseRestInfoPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
-                .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
-                .setHbaseRestReadOnly(
-                        Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
-                .setHbaseRestThreadMax(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
-                .setHbaseRestThreadMin(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
-                .build();
-    }
-
-    @Test
-    public void testHbaseNumRegionServers() {
-        assertEquals(Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_NUM_REGION_SERVERS_KEY)),
-                (int) hbaseRestLocalCluster.getNumRegionServers());
-    }
-
-    @Test
-    public void testMissingHbaseNumRegionsServers() {
-        exception.expect(IllegalArgumentException.class);
-        hbaseRestLocalCluster = new HbaseRestLocalCluster.Builder()
-                .setHbaseMasterPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_PORT_KEY)))
-                .setHbaseMasterInfoPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_INFO_PORT_KEY)))
-                .setHbaseRootDir(propertyParser.getProperty(ConfigVars.HBASE_ROOT_DIR_KEY))
-                .setZookeeperPort(Integer.parseInt(propertyParser.getProperty(ConfigVars.ZOOKEEPER_PORT_KEY)))
-                .setZookeeperConnectionString(propertyParser.getProperty(ConfigVars.ZOOKEEPER_CONNECTION_STRING_KEY))
-                .setZookeeperZnodeParent(propertyParser.getProperty(ConfigVars.HBASE_ZNODE_PARENT_KEY))
-                .setHbaseRestPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
-                .setHbaseRestInfoPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
-                .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
-                .setHbaseRestReadOnly(
-                        Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
-                .setHbaseRestThreadMax(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
-                .setHbaseRestThreadMin(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
-                .build();
-    }
-
-    @Test
-    public void testHbaseRootDir() {
-        assertEquals(propertyParser.getProperty(ConfigVars.HBASE_ROOT_DIR_KEY), hbaseRestLocalCluster.getHbaseRootDir());
-    }
-
-    @Test
-    public void testMissingHbaseRootDir() {
-        exception.expect(IllegalArgumentException.class);
-        hbaseRestLocalCluster = new HbaseRestLocalCluster.Builder()
-                .setHbaseMasterPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_PORT_KEY)))
-                .setHbaseMasterInfoPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_INFO_PORT_KEY)))
-                .setNumRegionServers(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_NUM_REGION_SERVERS_KEY)))
-                .setZookeeperPort(Integer.parseInt(propertyParser.getProperty(ConfigVars.ZOOKEEPER_PORT_KEY)))
-                .setZookeeperConnectionString(propertyParser.getProperty(ConfigVars.ZOOKEEPER_CONNECTION_STRING_KEY))
-                .setZookeeperZnodeParent(propertyParser.getProperty(ConfigVars.HBASE_ZNODE_PARENT_KEY))
-                .setHbaseRestPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
-                .setHbaseRestInfoPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
-                .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
-                .setHbaseRestReadOnly(
-                        Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
-                .setHbaseRestThreadMax(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
-                .setHbaseRestThreadMin(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
-                .build();
-    }
-
-    @Test
-    public void testZookeeperPort() {
-        assertEquals(Integer.parseInt(propertyParser.getProperty(ConfigVars.ZOOKEEPER_PORT_KEY)),
-                (int) hbaseRestLocalCluster.getZookeeperPort());
-    }
-
-    @Test
-    public void testMissingZookeeperPort() {
-        exception.expect(IllegalArgumentException.class);
-        hbaseRestLocalCluster = new HbaseRestLocalCluster.Builder()
-                .setHbaseMasterPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_PORT_KEY)))
-                .setHbaseMasterInfoPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_INFO_PORT_KEY)))
-                .setNumRegionServers(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_NUM_REGION_SERVERS_KEY)))
-                .setHbaseRootDir(propertyParser.getProperty(ConfigVars.HBASE_ROOT_DIR_KEY))
-                .setZookeeperConnectionString(propertyParser.getProperty(ConfigVars.ZOOKEEPER_CONNECTION_STRING_KEY))
-                .setZookeeperZnodeParent(propertyParser.getProperty(ConfigVars.HBASE_ZNODE_PARENT_KEY))
-                .setHbaseRestPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
-                .setHbaseRestInfoPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
-                .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
-                .setHbaseRestReadOnly(
-                        Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
-                .setHbaseRestThreadMax(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
-                .setHbaseRestThreadMin(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
-                .build();
-    }
-
-    @Test
-    public void testZookeeperConnectionString() {
-        assertEquals(propertyParser.getProperty(ConfigVars.ZOOKEEPER_CONNECTION_STRING_KEY),
-                hbaseRestLocalCluster.getZookeeperConnectionString());
-    }
-
-    @Test
-    public void testMissingZookeeperConnectionString() {
-        exception.expect(IllegalArgumentException.class);
-        hbaseRestLocalCluster = new HbaseRestLocalCluster.Builder()
-                .setHbaseMasterPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_PORT_KEY)))
-                .setHbaseMasterInfoPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_INFO_PORT_KEY)))
-                .setNumRegionServers(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_NUM_REGION_SERVERS_KEY)))
-                .setHbaseRootDir(propertyParser.getProperty(ConfigVars.HBASE_ROOT_DIR_KEY))
-                .setZookeeperPort(Integer.parseInt(propertyParser.getProperty(ConfigVars.ZOOKEEPER_PORT_KEY)))
-                .setZookeeperZnodeParent(propertyParser.getProperty(ConfigVars.HBASE_ZNODE_PARENT_KEY))
-                .setHbaseRestPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
-                .setHbaseRestInfoPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
-                .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
-                .setHbaseRestReadOnly(
-                        Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
-                .setHbaseRestThreadMax(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
-                .setHbaseRestThreadMin(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
-                .build();
-    }
-
-    @Test
-    public void testZookeeperZnodeParent() {
-        assertEquals(propertyParser.getProperty(ConfigVars.HBASE_ZNODE_PARENT_KEY),
-                hbaseRestLocalCluster.getZookeeperZnodeParent());
-    }
-
-    @Test
-    public void testMissingZookeeperZnodeParent() {
-        exception.expect(IllegalArgumentException.class);
-        hbaseRestLocalCluster = new HbaseRestLocalCluster.Builder()
-                .setHbaseMasterPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_PORT_KEY)))
-                .setHbaseMasterInfoPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_INFO_PORT_KEY)))
-                .setNumRegionServers(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_NUM_REGION_SERVERS_KEY)))
-                .setHbaseRootDir(propertyParser.getProperty(ConfigVars.HBASE_ROOT_DIR_KEY))
-                .setZookeeperPort(Integer.parseInt(propertyParser.getProperty(ConfigVars.ZOOKEEPER_PORT_KEY)))
-                .setZookeeperConnectionString(propertyParser.getProperty(ConfigVars.ZOOKEEPER_CONNECTION_STRING_KEY))
-                .setHbaseRestPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
-                .setHbaseRestInfoPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
-                .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
-                .setHbaseRestReadOnly(
-                        Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
-                .setHbaseRestThreadMax(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
-                .setHbaseRestThreadMin(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
-                .build();
-    }
-
-
-    @Test
-    public void testHbaseRestPort() {
-        assertEquals(Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)),
-                (int) hbaseRestLocalCluster.getHbaseRestPort());
-    }
-
-    @Test
-    public void testMissingHbaseRestPort() {
-        exception.expect(IllegalArgumentException.class);
-        hbaseRestLocalCluster = new HbaseRestLocalCluster.Builder()
-                .setHbaseMasterPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_PORT_KEY)))
-                .setHbaseMasterInfoPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_INFO_PORT_KEY)))
-                .setNumRegionServers(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_NUM_REGION_SERVERS_KEY)))
-                .setHbaseRootDir(propertyParser.getProperty(ConfigVars.HBASE_ROOT_DIR_KEY))
-                .setZookeeperPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.ZOOKEEPER_PORT_KEY)))
-                .setZookeeperConnectionString(propertyParser.getProperty(ConfigVars.ZOOKEEPER_CONNECTION_STRING_KEY))
-                .setZookeeperZnodeParent(propertyParser.getProperty(ConfigVars.HBASE_ZNODE_PARENT_KEY))
-                .setHbaseRestInfoPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
-                .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
-                .setHbaseRestReadOnly(
-                        Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
-                .setHbaseRestThreadMax(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
-                .setHbaseRestThreadMin(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
-                .build();
-    }
-
-    @Test
-    public void testHbaseRestInfoPort() {
-        assertEquals(Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)),
-                (int) hbaseRestLocalCluster.getHbaseRestInfoPort());
-    }
-
-    @Test
-    public void testHbaseRestHost() {
-        assertEquals(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY),
-                hbaseRestLocalCluster.getHbaseRestHost());
-    }
-
-    @Test
-    public void testHbaseRestReadOnly() {
-        assertEquals(Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)),
-                hbaseRestLocalCluster.getHbaseRestReadOnly());
-    }
-
-    @Test
-    public void testHbaseRestThreadMax() {
-        assertEquals(Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)),
-                hbaseRestLocalCluster.getHbaseRestThreadMax());
-    }
-
-    @Test
-    public void testHbaseRestThreadMin() {
-        assertEquals(Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)),
-                hbaseRestLocalCluster.getHbaseRestThreadMin());
-    }
-
-}
-
