+++ hadoop-mini-clusters-knox/src/test/java/com/github/sakserv/minicluster/impl/KnoxLocalClusterIntegrationTest.java
@@ -69,7 +69,6 @@ public class KnoxLocalClusterIntegrationTest {
         }
     }
 
-    private static HbaseRestLocalCluster hbaseRestLocalCluster;
     private static HbaseLocalCluster hbaseLocalCluster;
     private static ZookeeperLocalCluster zookeeperLocalCluster;
     private static HdfsLocalCluster dfsCluster;
@@ -99,33 +98,19 @@ public class KnoxLocalClusterIntegrationTest {
                 .setHbaseWalReplicationEnabled(
                         Boolean.parseBoolean(propertyParser.getProperty(ConfigVars.HBASE_WAL_REPLICATION_ENABLED_KEY)))
                 .setHbaseConfiguration(new Configuration())
+                .activeRestGateway()
+                    .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
+                    .setHbaseRestPort(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
+                    .setHbaseRestReadOnly(
+                            Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
+                    .setHbaseRestThreadMax(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
+                    .setHbaseRestThreadMin(
+                            Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
+                    .build()
                 .build();
         hbaseLocalCluster.start();
-
-        hbaseRestLocalCluster = new HbaseRestLocalCluster.Builder()
-                .setHbaseMasterPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_PORT_KEY)))
-                .setHbaseMasterInfoPort(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_MASTER_INFO_PORT_KEY)))
-                .setNumRegionServers(
-                        Integer.parseInt(propertyParser.getProperty(ConfigVars.HBASE_NUM_REGION_SERVERS_KEY)))
-                .setHbaseRootDir(propertyParser.getProperty(ConfigVars.HBASE_ROOT_DIR_KEY))
-                .setZookeeperPort(Integer.parseInt(propertyParser.getProperty(ConfigVars.ZOOKEEPER_PORT_KEY)))
-                .setZookeeperConnectionString(propertyParser.getProperty(ConfigVars.ZOOKEEPER_CONNECTION_STRING_KEY))
-                .setZookeeperZnodeParent(propertyParser.getProperty(ConfigVars.HBASE_ZNODE_PARENT_KEY))
-                .setHbaseRestPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_PORT_KEY)))
-                .setHbaseRestInfoPort(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_INFO_PORT_KEY)))
-                .setHbaseRestHost(propertyParser.getProperty(ConfigVars.HBASE_REST_HOST_KEY))
-                .setHbaseRestReadOnly(
-                        Boolean.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_READONLY_KEY)))
-                .setHbaseRestThreadMax(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMAX_KEY)))
-                .setHbaseRestThreadMin(
-                        Integer.valueOf(propertyParser.getProperty(ConfigVars.HBASE_REST_THREADMIN_KEY)))
-                .build();
-        hbaseRestLocalCluster.start();
 
         // We need HDFS/WEBHDFS for Knox
         dfsCluster = new HdfsLocalCluster.Builder()
@@ -177,7 +162,6 @@ public class KnoxLocalClusterIntegrationTest {
 
     @AfterClass
     public static void tearDown() throws Exception {
-        hbaseRestLocalCluster.stop();
         hbaseLocalCluster.stop();
         zookeeperLocalCluster.stop();
         dfsCluster.stop();
